Fix security issues with api returning unwanted field

+++ server/controllers/usersTourneysController.js
@@ -1,8 +1,11 @@
-const { UsersTourneys, User, Tourney } = require('../models');
+// server/controllers/usersTourneysController.js
+const { UsersTourneys, User, Tourney, Team, Role } = require('../models');
+const { Op } = require('sequelize');
 
 // Ajouter un utilisateur à un tournoi
 exports.addUserToTourney = async (req, res) => {
-    const { userId, tourneyId } = req.body;
+    const { userId } = req.body;
+    const { tourneyId } = req.params;
 
     try {
         const userTourney = await UsersTourneys.create({ userId, tourneyId });
@@ -13,30 +16,30 @@ exports.addUserToTourney = async (req, res) => {
     }
 };
 
-// Récupérer tous les tournois auxquels un utilisateur participe
-exports.getTourneysByUser = async (req, res) => {
-    const { userId } = req.params;
-
-    try {
-        const tourneys = await UsersTourneys.findAll({
-            where: { userId },
-            include: [Tourney],
-        });
-        res.json(tourneys);
-    } catch (error) {
-        console.error('Erreur lors de la récupération des tournois pour l’utilisateur:', error);
-        res.status(500).json({ message: 'Erreur serveur', error });
-    }
-};
-
-// Récupérer tous les utilisateurs d'un tournoi
+// Récupérer tous les utilisateurs d'un tournoi (hors admin)
 exports.getUsersByTourney = async (req, res) => {
     const { tourneyId } = req.params;
 
     try {
-        const users = await UsersTourneys.findAll({
-            where: { tourneyId },
-            include: [User],
+        const users = await User.findAll({
+            attributes: ['id', 'name', 'roleId', 'teamId', 'Team'],
+            include: [
+                {
+                    model: Tourney,
+                    as: 'tourneys',
+                    where: { id: tourneyId },
+                    attributes: []
+                },
+                {
+                    model: Role,
+                    where: { name: { [Op.ne]: 'admin' } },
+                    attributes: []
+                },
+                {
+                    model: Team,
+                    attributes: ['teamName', 'type']
+                },
+            ],
         });
         res.json(users);
     } catch (error) {
@@ -45,14 +48,29 @@ exports.getUsersByTourney = async (req, res) => {
     }
 };
 
-// Supprimer la participation d'un utilisateur à un tournoi
+// Supprimer la participation d'un utilisateur à un tournoi (protection des admins)
 exports.removeUserFromTourney = async (req, res) => {
     const { userId, tourneyId } = req.params;
 
     try {
+        // Récupérer l'utilisateur
+        const user = await User.findByPk(userId);
+
+        if (!user) {
+            return res.status(404).json({ message: "Utilisateur non trouvé." });
+        }
+
+        // Vérifier que l'utilisateur n'est pas un Admin
+        const role = await Role.findByPk(user.roleId);
+        if (role.name === 'admin') {
+            return res.status(403).json({ message: "Impossible de supprimer un Admin du tournoi." });
+        }
+
+        // Supprimer la participation de l'utilisateur au tournoi
         await UsersTourneys.destroy({
             where: { userId, tourneyId },
         });
+
         res.status(204).send();
     } catch (error) {
         console.error('Erreur lors de la suppression de l’utilisateur du tournoi:', error);
@@ -60,58 +78,65 @@ exports.removeUserFromTourney = async (req, res) => {
     }
 };
 
+// Récupérer les utilisateurs non assignés à une équipe pour un tournoi spécifique (Guest users)
 exports.getUnassignedUsersByTourney = async (req, res) => {
     const { tourneyId } = req.params;
     try {
-      const unassignedUsers = await User.findAll({
-        where: {
-          teamId: null,
-          roleId: 4,
-          '$Tourneys.id$': tourneyId,
-        },
-        include: [
-          {
-            model: Tourney,
-            as: 'tourneys',
-            where: { id: tourneyId },
-          },
-        ],
-      });
-      res.json(unassignedUsers);
-    } catch (error) {
-      res.status(500).json({ error: 'Erreur lors de la récupération des utilisateurs non assignés.' });
-    }
-  };
-
-  // Récupérer les informations d'un utilisateur dans un tournoi
-exports.getUserInfoByTourney = async (req, res) => {
-    const { tourneyId, idUser } = req.params;
-
-    try {
-        // Vérifier que l'utilisateur participe bien au tournoi via UsersTourneys
-        const userInTourney = await UsersTourneys.findOne({
-            where: { userId: idUser, tourneyId },
+        const unassignedUsers = await User.findAll({
+            where: {
+                teamId: null,
+                roleId: 4, // Assurez-vous que 4 correspond bien à 'guest'
+            },
+            attributes: ['id', 'name', 'roleId', 'teamId'],
             include: [
                 {
-                    model: User,
-                    attributes: ['id', 'name', 'email'], // Informations générales de l'utilisateur
+                    model: Tourney,
+                    as: 'tourneys',
+                    where: { id: tourneyId },
+                    attributes: []
+                },
+            ],
+        });
+        res.json(unassignedUsers);
+    } catch (error) {
+        console.error('Erreur lors de la récupération des utilisateurs non assignés :', error);
+        res.status(500).json({ error: 'Erreur lors de la récupération des utilisateurs non assignés.' });
+    }
+};
+
+// Récupérer les informations d'un utilisateur dans un tournoi, y compris son équipe
+exports.getUserInfoByTourney = async (req, res) => {
+    const { tourneyId, userId } = req.params;
+
+    try {
+        const user = await User.findOne({
+            where: { id: userId },
+            attributes: ['id', 'name', 'roleId', 'teamId', 'Team'],
+            include: [
+                {
+                    model: Tourney,
+                    as: 'tourneys',
+                    where: { id: tourneyId },
+                    attributes: []
                 },
                 {
                     model: Team,
-                    attributes: ['teamName', 'type'], // Informations sur l'équipe (si applicable)
-                }
-            ]
+                    attributes: ['teamName', 'type']
+                },
+                {
+                    model: Role,
+                    attributes: ['name']
+                },
+            ],
         });
 
-        if (!userInTourney) {
+        if (!user) {
             return res.status(404).json({ message: "L'utilisateur ne participe pas à ce tournoi." });
         }
 
-        // Renvoyer les informations de l'utilisateur, y compris son rôle et son équipe
-        res.status(200).json(userInTourney);
+        res.status(200).json(user);
     } catch (error) {
         console.error('Erreur lors de la récupération des informations de l\'utilisateur dans le tournoi:', error);
         res.status(500).json({ message: 'Erreur serveur', error });
     }
 };
-
